move the seed decorator to utils
